backend/services/admin_service/requirements.txt
Flask
grpcio
grpcio-tools
requests
python-dotenv
flask-cors
